Refactor LOCAL pipeline implementation
